refactor: move babel.js into utils/
To avoid confusion, we should keep things that aren't scripts
corresponding to subcommands that you can run with "liferay-npm-scripts
subcommand" out of "src/scripts/".
